requirements.txt
flask
requests
mitmproxy
scapy
keyboard
pyautogui
python-dotenv
